validation-pipeline/blr-final/requirements1.txt
annotated-types==0.7.0
anyio==4.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
docstring_parser==0.17.0
elevenlabs==2.8.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
firebase_admin==7.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-resource-manager==1.14.2
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googlemaps==4.10.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
numpy==2.3.2
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
rich-toolkit==0.14.8
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.33.2
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vertexai==1.71.1
watchfiles==1.1.0
websockets==15.0.1
